src/requirements_ubuntu.txt
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.0.0
pathlib==1.0.1
py-tlsh==4.5.0
pydantic==1.8.1
PyMySQL==1.0.2
smmap==4.0.0
SQLAlchemy==1.4.13
typing-extensions==3.7.4.3
schedule == 1.1.0
JPype1==1.3.0